refactor(ProductApiController): 코드 개선
- 롬복 적용
- 변수명 수정
